Refactor connector and remove local fungible_token implementation
