--- evm-fungible-token/src/connector/mod.rs
@@ -1,4 +1,0 @@
-#![allow(dead_code, unused_imports, unused_variables)]
-pub mod deposit_event;
-pub mod prover;
-pub mod withdraw_event;
