--- evm-fungible-token/src/fungible_token/fungible_token_metadata.rs
@@ -1,22 +1,0 @@
-use super::*;
-use near_sdk::serde::Serialize;
-
-#[derive(Serialize, BorshDeserialize, BorshSerialize, Clone)]
-#[serde(crate = "near_sdk::serde")]
-pub struct FungibleTokenMetadata {
-    pub version: String,
-    pub name: String,
-    pub symbol: String,
-    pub reference: String,
-    pub decimals: u8,
-}
-
-pub trait FungibleTokenMetadataProvider {
-    fn ft_metadata(&self) -> FungibleTokenMetadata;
-}
-
-impl FungibleTokenMetadataProvider for FungibleToken {
-    fn ft_metadata(&self) -> FungibleTokenMetadata {
-        self.metadata.clone()
-    }
-}
